Add additional asp identity setting - server side identity store
This was added to resolve a possible future problem - since all browsers
have cookie size limits, a cookie too large will cause session break.
Size of a asp identity cookie grows with each role / claim added to user,
therefore it can cause errors in the future if claims granularity will be large.

This new settings allow to switch asp cookie to a version that
will only point to a in-memory identity storage, which will enable any possible number of
claims and roles will to be handled.