Issue #8 GitLab - Cancel button is now only enabled when needed.
Add IsCancelButtonDisabled() method which checks
couple of parameters before deciding if cancel button
should be enabled.

Implement Save(TItem titem) in SpareParts.razor so that
StateHasChanged is called after something is edited in
the table. This is needed for proper rendering of the
cancel button.

Cancel button does not call a database -> fixed in one
of the previous commits.

Add call of InvokeSaveChanges() in ScanAppTable.razor
after editing dialog is submitted successfuly.