#7 - Add validation for "From" and "To" fields in filtering dialog.
Add logic in (...)FilterDialogFieldGenerator.razor so that
user cannot submit filtering when:
- From is larger than To
- user enters negative numbers (depends on ScanAppTable Options)

Add functionality that marks fields which contains errors,
displays one error message when it concerns two fields
instead of displaying two messages. Also, opens the panels which
contain fields with errors so that users is made aware of the problem.