#23 Editing in a filtered state is now possible without resetting the filter
Change how ScanAppTable handles collections
on the inside. Now there are two collections:

- ShownItems (the collection visible to user in the table)
- Items (the original collection provided by the developer)

ShownItems are subject to editing, filtering and
grouping while Items are only subject to editing.

Add UpdateTable(List<TItem> items) public method so
that updating/changing the collection visible to user
is much easier. The method changes both ShownItems
and Items. Keep it mind that it also disables filtering
and grouping state to ensure no errors are introduced.

Add private methods used for determining if two items
are the same. In contrary to methods provided with .Net
these methods work on cloned objects, not by reference.
These methods are used to ensure proper editing of both
ShownItems and Items.