#19 - Change how ScanAppTableOptions operate.
Now ShowToolBar is controlled automatically.
If one option ie. AllowFiltering is enabled, ShowToolBar
is enabled automatically as well. On the other hand,
if both filtering and grouping is disabled, toolbar
is disabled as well.

Moreover, if no Options are passed to the ScanAppTable,
all operations will be enabled by default.

Using default constructor of ScanAppTableOptions
without any parameters will result in all operations
to be disabled.