#26 - add validation functionality for the ScanAppTable / negative numbers not allowed
The Validators (for example FluentValidationWrapper or
a validator that inherits from AbstractValidator<>) can
now be used. They can be defined in ColumnConfig
class.

Once defined, they are implemented in filter and edit
fields for each property. The ok button for both editing
and filtering can only be clicked if the validation gives a
positive result.