Change how filter dialog is generated. Now works with all datatypes.
The fields are dynamically generated using RenderTreeBuilder.
Validation is integrated with ColumnConfig.cs. Some
fields are made nullable for enhanced user experienced. That
means that some validators are cast to a nullable if possible
so that they work properly with nullable fields. Additionally
introduced some unit tests for new filtering operation for
decimals.