Implement "Add" feature to the ScanAppTable
The new items can now be added to the table.
Feature can be enabled using ScanAppTableOptions.
If user wants to have a default, not editable value in
adding dialog, a Default property can be defined
in ColumnConfig.

After the item is added, the call is made to
ItemAdded event callback. User can access
the last added item through RecentlyAddedItem
property of the ScanAppTable.