Fluent validation wrapper has now settable null reaction.
If null is valid in given context, then when instantiating
wrapper value 'allowNull' should be set to  true (default). This is caused by FluidValidation way of handling context - Not a bug, but a feature, said the author. Thanks to this, workaround was needed.
Otherwise we could not use IValidators but would need to use
Property validators, that are somewhat limited.