Add unique constraint exception handler.
It is derived from DbUpdateExceptionHndler for a purpose - Since handlers are a conept of
Mediatr pipeline, execution order is defined by
pipeline handler.
This is explained here:
https://github.com/jbogard/MediatR/wiki#handlers-and-actions-priority-execution

So to force usage of UniqueConstr. handler, since it is handling a derived extension,
I choose to override priority by deriving from more general handler.

This is the way it should be handled from now on.