Departure plans for date range query WiP:
Added a query to check if any of saved recurring plans is in range
of main query - if so, it should be download. This unfortunately requires
at least two calls to database, given that pattern occ. calculations are
to complicated (?) to drop them into SQL server easily without need for
scalar functions, if that is even possible.