Add new way of mapping Departure plan entities to models
New way is respecting entity references.
Since data provided is not creating separate objects for each entity and exception to
it, therefore similar way is taken when creating models - this is planned.

In other words, when returning main departure plan and other departure plan that is
an exception to it, there will be only two objects created - main departure plan and
exception. Exception will only reference main object, it will not hold a
completely new, separate model of main plan to which it is an exception.